requirements.txt
redis==5.0.1
redisai==1.0.0
langgraph==0.2.0
langchain==0.2.0
langchain-openai==0.1.0
numpy==1.24.3
torch==2.0.1
transformers==4.30.2
python-dotenv==1.0.0 